K-Foldable String
You are given a binary string S with length N and an integer K, which is a divisor of N. A string is said to be K-foldable if it can be changed to a string with length K by repeating the following process without any collisions (defined below):

Select the prefix of the current string S with length 2K. (Note that as long as the length of S is greater than K, this prefix always exists.)
For each i (1≤i≤K), check if the 2K−i+1-th character of S is equal to the i-th character of S ― if they are not equal, there is a collision and the process is invalid.
Erase the prefix of S with length K.
Your goal is to reorder the characters of S (possibly leaving this string unchanged) in such a way that the resulting string S is a K-foldable string. Find a way to do that or determine that it is impossible. If there are multiple solutions, find the lexicographically smallest rearranged string which is K-foldable.