What Is This, a Crossover Episode
You are given an integer D. You have to find a string S such that the following conditions are satisfied:

S is non-empty and its length does not exceed 7⋅D.
S contains only lowercase English letters.
The number of distinct substrings in S minus the number of palindromic substrings in S equals D. Here, when counting palindromic substrings, a substring that occurs multiple times should be counted multiple times.
It can be proved that a solution always exists under the given constraints.

Notes:

A string B is a substring of a string A if B can be obtained from A by deleting several (possibly none or all) characters from the beginning and several (possibly none or all) characters from the end.
A palindrome is a string that does not change when it is reversed. E.g. strings "z", "aba" and "dammitimmad" are palindromes, but strings "ab" and "bojack" are not.
A palindromic substring is a substring which is also a palindrome.