XOR Engine
A sophomore Computer Science student is frustrated with boring college lectures. Professor X agreed to give him some questions; if the student answers all questions correctly, then minimum attendance criteria will not apply to him.

Professor X chooses a sequence A1,A2,…,AN and asks Q queries. In each query, the student is given an integer P; he has to construct a sequence B1,B2,…,BN, where P⊕Ai=Bi for each valid i (⊕ denotes bitwise XOR), and then he has to find the number of elements of this sequence which have an even number of 1-s in the binary representation and the number of elements with an odd number of 1-s in the binary representation. Help him answer the queries.