Convert the String
Vivek was quite bored with the lockdown, so he came up with an interesting task. He successfully completed this task and now, he would like you to solve it.

You are given two strings A and B, each with length N. Let's index the characters in each string from 0 ― for each i (0≤i<N), the i+1-th characters of A and B are denoted by Ai and Bi respectively. You should convert A to B by performing operations of the following type:

Choose a subset S of the set {0,1,…,N−1}.
Let c be the alphabetically smallest character among Ax for all x∈S.
For each x∈S, replace Ax by c.
You should find the smallest necessary number of operations or report that it is impossible to convert A to B. If it is possible, you also need to find one way to convert A to B using this smallest number of operations. If there are multiple solutions, you may find any one.